Move HSQLDB checkpoint/shutdown to a bean specific to the legacy profile
