+++ airsonic-main/src/main/java/org/airsonic/player/dao/DaoHelper.java
@@ -53,7 +53,7 @@ public interface DaoHelper {
      * Database checkpoints will make sure that the database is written on the disk
      * and optimize on-disk storage.
      *
-     * @return true if the checkpoint succeeded
+     * @return true if the checkpoint succeeded, false otherwise
      */
-    public boolean checkpoint();
+    public default boolean checkpoint() { return false; }
 }
