+++ airsonic-main/src/main/resources/applicationContext-db-legacy.xml
@@ -6,11 +6,17 @@
 
     <bean id="dataSource"
           class="org.springframework.jdbc.datasource.DriverManagerDataSource">
-
         <property name="driverClassName" value="org.hsqldb.jdbcDriver" />
         <property name="url"
                   value="#{T(org.airsonic.player.service.SettingsService).defaultJDBCUrl}" />
         <property name="username" value="sa" />
         <property name="password" value="" />
     </bean>
+
+    <!-- Overwrite the GenericDaoHelper bean defined in applicationContext-db.xml -->
+    <!-- This bean is specific to the legacy embedded HSQLDB database -->
+    <bean id="daoHelper" class="org.airsonic.player.dao.LegacyHsqlDaoHelper">
+        <constructor-arg name="dataSource" ref="dataSource" />
+    </bean>
+
 </beans>
