+++ airsonic-main/src/main/resources/applicationContext-db.xml
@@ -3,12 +3,6 @@
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:tx="http://www.springframework.org/schema/tx"
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx.xsd">
 
-    <import resource="applicationContext-db-jndi.xml" />
-    <import resource="applicationContext-db-embed.xml" />
-    <import resource="applicationContext-db-legacy.xml" />
-
-    <tx:annotation-driven/>
-
     <bean id="transactionManager" class="org.springframework.jdbc.datasource.DataSourceTransactionManager">
         <constructor-arg ref="dataSource"/>
     </bean>
@@ -16,6 +10,12 @@
     <bean id="daoHelper" class="org.airsonic.player.dao.GenericDaoHelper">
         <constructor-arg name="dataSource" ref="dataSource" />
     </bean>
+
+    <import resource="applicationContext-db-jndi.xml" />
+    <import resource="applicationContext-db-embed.xml" />
+    <import resource="applicationContext-db-legacy.xml" />
+
+    <tx:annotation-driven/>
 
     <bean id="rollbackFile" class="java.io.File">
         <constructor-arg type="java.io.File" index="0" value="#{T(org.airsonic.player.service.SettingsService).airsonicHome}" />
